Split deposits html page code by table
